move Letter to new file, add Mailbox alias and update usages
