+++ src/actor/mod.rs
@@ -4,6 +4,10 @@
 
 pub mod actor;
 pub mod address;
+pub mod letter;
 
 pub use actor::*;
 pub use address::*;
+pub use letter::*;
+
+pub type Mailbox = crossbeam_channel::Sender<Box<dyn crate::message::Message>>;
